Fix some logic flaws exposed during recovery
1. Backup $HOME
1. Install fresh OS
1. Restore $HOME

Applications outside of $HOME have not been installed yet, but there are
assumptions in these configuration files that make assumptions and
expect certain things to be in place.

Any calls to, or dependancies on, utilities outside of $HOME should have
the appropriate gating logic or a fallback that will be substituted when
the preferred is not available.

e.g.
- Explicit references, like `/usr/bin/which`, remove ambiguity
- `greadlink` should have a fallback
- Gating logic could disable blocks of code or modules modules instead
  of trying to handle all cases... just test required assumptions and if
  they aren't met, then no-op.